==> scripts/2_upload_data.ipynb <==
import pandas as pd
import time

# Loads all the csv data into dataframes
sesiones_df = pd.read_csv("data/processed/sesiones.csv")
votaciones_df = pd.read_csv("data/processed/votaciones.csv")
grupos_parlamentarios_df = pd.read_csv("data/processed/grupos_parlamentarios.csv")
diputados_df = pd.read_csv("data/processed/diputados.csv")
votos_df = pd.read_csv("data/processed/votos.csv")
proponentes_df = pd.read_csv("data/processed/proponentes.csv")
categorias_df = pd.read_csv("categorias.csv")
subcategorias_df = pd.read_csv("subcategorias.csv")
votaciones_subcategorias_df = pd.read_csv("data/processed/votaciones_subcategorias.csv")

# Uploads the dataframes to the supabase instance
from supabase import create_client, Client

# read key from .env.local file NEXT_PUBLIC_SUPABASE_URL
with open("../.env.local", "r") as f:
    lines = f.readlines()
    url = lines[0].split("=")[1].strip()
with open("data/supabase_key.txt", "r") as f:
    lines = f.readlines()
    key = lines[0].strip()

supabase: Client = create_client(url, key)



# Uploads the dataframes to the supabase instance
print("Deleting...")
table_pks = [
    ("proponentes", "id_proponente"),
    ("votaciones_subcategorias", "id_votacion_subcategoria"),
    ("subcategorias", "id_subcategoria"),
    ("categorias", "id_categoria"),
    ("votos", "id_voto"),
    ("diputados", "id_diputado"),
    ("grupos_parlamentarios", "id_grupo"),
    ("votaciones", "id_votacion"),
    ("sesiones", "id_sesion"),
]
for table, pk in table_pks:
    print(f"Deleting {table}...")
    # Delete all rows in the table with id >= 0
    try:
        supabase.table(table).delete().gte(pk, 0).execute()
    except Exception as e:
        print(f"Error deleting {table}: {e}")
        continue
    time.sleep(5)  # Sleep for 1 second to avoid rate limiting
    print(f"Deleted {table}")


print("Inserting...")
supabase.table("sesiones").insert(sesiones_df.to_dict(orient="records")).execute()
time.sleep(1)
supabase.table("votaciones").insert(votaciones_df.to_dict(orient="records")).execute()
time.sleep(1)
supabase.table("grupos_parlamentarios").insert(grupos_parlamentarios_df.to_dict(orient="records")).execute()
time.sleep(1)
supabase.table("diputados").insert(diputados_df.to_dict(orient="records")).execute()
time.sleep(1)
# Define a batch size
batch_size = 10000

# Split the DataFrame into smaller chunks
for i in range(0, len(votos_df), batch_size):
    print(f"Inserting batch {i // batch_size + 1} of {len(votos_df) // batch_size + 1}")
    batch = votos_df.iloc[i:i + batch_size]
    supabase.table("votos").insert(batch.to_dict(orient="records")).execute()
    time.sleep(5)  # Optional: Add a delay between batches to avoid overwhelming the server

supabase.table("proponentes").insert(proponentes_df.to_dict(orient="records")).execute()
time.sleep(1)
supabase.table("categorias").insert(categorias_df.to_dict(orient="records")).execute()
time.sleep(1)
supabase.table("subcategorias").insert(subcategorias_df.to_dict(orient="records")).execute()
time.sleep(1)
supabase.table("votaciones_subcategorias").insert(votaciones_subcategorias_df.to_dict(orient="records")).execute()
